django-assignment-1-D1183-Mustafa/requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.3.0
click==8.1.2
Django==4.0.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest-django==4.5.2
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
